refactor: extract status and adoption
